# perf(ext/request): optimize Request constructor